+++ src/assets/styles/_profile.scss
@@ -1,0 +1,22 @@
+.profile-editor-input {
+	text-align: left;
+	border: none;
+}
+
+.profile_editor_form p {
+	width: 50%;
+	float: left;
+}
+
+.profile_editor_form p {
+	width: 50%;
+	float: left;
+}
+
+.profile_editor_form p:nth-child(2n + 2) {
+	padding-left: 5px;
+}
+
+.profile_editor_form p:nth-child(1n + 1) {
+	padding-right: 5px;
+}
