+++ src/assets/styles/index.scss
@@ -2,3 +2,4 @@
 @import "login";
 @import "ant-override";
 @import "modals";
+@import "profile";
